refactor: dialog 오타 수정 및 DialogTypes types/ 폴더로 분리
