+++ src/types/ui/dialog.ts
@@ -1,0 +1,6 @@
+export type DialogTypes =
+  | 'example'
+  | 'changeJob'
+  | 'submitAnswer'
+  | 'expiredSession'
+  | 'continueWriting';
